got config toggles to work
Had to figure out a system to make them all work at the same time.

added them to a boolean list, each toggle adds some condition to a list.

The list is iterated through, if any of the toggles are false, it breaks. They must all be true for the render to trigger